+++ .travis.yml
@@ -4,4 +4,5 @@ node_js:
 before_install:
   - currentfolder=${PWD##*/}
   - if [ "$currentfolder" != 'generator-git' ]; then cd .. && eval "mv $currentfolder generator-git" && cd generator-git; fi
-
+script:
+  - npm run test
